code refactoring:
added constants for controllers
extended admin and booking controllers
renamed some dtos to start with non-verb
added javaDoc for services
used facade pattern for user service
