+++ src/main/java/com/example/cdr/eventsmanagementsystem/Constants/ServiceControllerConstants.java
@@ -1,0 +1,12 @@
+package com.example.cdr.eventsmanagementsystem.Constants;
+
+public class ServiceControllerConstants {
+
+    public static final String SERVICE_BASE_URL = "/v1/services";
+    public static final String CREATE_SERVICE_URL ="/create";
+    public static final String UPDATE_SERVICE_AVAILABILITY = "/{serviceId}/availability";
+    public static final String GET_SERVICE_BOOKINGS_URL = "/bookings";
+    public static final String ACCEPT_OR_REJECT_SERVICE_BOOKING_URL = "/bookings/{bookingId}/status";
+    public static final String CANCEL_SERVICE_BOOKING_URL = "/bookings/{bookingId}/cancel";
+
+}
