+++ src/main/java/com/example/cdr/eventsmanagementsystem/Constants/VenueControllerConstants.java
@@ -1,0 +1,10 @@
+package com.example.cdr.eventsmanagementsystem.Constants;
+
+public class VenueControllerConstants {
+    public static final String VENUE_BASE_URL = "/v1/venues";
+    public static final String CREATE_VENUE_URL = "/create";
+    public static final String UPDATE_VENUE_URL = "/{venueId}";
+    public static final String DELETE_VENUE_URL = "/{venueId}";
+    public static final String GET_VENUE_BOOKINGS_URL = "/bookings";
+    public static final String CANCEL_VENUE_BOOKING_URL = "/bookings/{bookingId}/cancel";
+}
