+++ src/main/java/com/example/cdr/eventsmanagementsystem/Controller/BookingController/BookingController.java
@@ -1,0 +1,9 @@
+package com.example.cdr.eventsmanagementsystem.Controller.BookingController;
+
+import com.example.cdr.eventsmanagementsystem.Constants.BookingControllerConstants;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@RequestMapping(BookingControllerConstants.BOOKING_BASE_URL)
+public abstract class BookingController {
+
+}
