+++ src/main/java/com/example/cdr/eventsmanagementsystem/Controller/ServiceController.java
@@ -1,12 +1,12 @@
 package com.example.cdr.eventsmanagementsystem.Controller;
 
-import com.example.cdr.eventsmanagementsystem.Constants.ControllerConstants;
 import com.example.cdr.eventsmanagementsystem.Constants.RoleConstants;
+import com.example.cdr.eventsmanagementsystem.Constants.ServiceControllerConstants;
 import com.example.cdr.eventsmanagementsystem.DTO.Booking.Response.BookingDetailsResponse;
 import com.example.cdr.eventsmanagementsystem.DTO.Service.ServicesDTO;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.Booking;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.BookingStatus;
-import com.example.cdr.eventsmanagementsystem.Service.Service.ServicesService;
+import com.example.cdr.eventsmanagementsystem.Service.Service.ServicesServiceInterface;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -16,18 +16,17 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-import java.nio.file.AccessDeniedException;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping(ControllerConstants.SERVICE_BASE_URL)
+@RequestMapping(ServiceControllerConstants.SERVICE_BASE_URL)
 @Tag(name = "Service", description = "Service provider APIs for managing services and bookings")
 public class ServiceController {
-    private final ServicesService servicesService;
+    private final ServicesServiceInterface servicesService;
 
     @Operation(summary = "Add a new service")
     @PreAuthorize("hasRole('" + RoleConstants.SERVICE_PROVIDER_ROLE + "')")
-    @PostMapping
+    @PostMapping(ServiceControllerConstants.CREATE_SERVICE_URL)
     public ResponseEntity<ServicesDTO> addService(@Valid @RequestBody ServicesDTO dto) {
         ServicesDTO servicesDTO = servicesService.addService(dto);
         return ResponseEntity.ok(servicesDTO);
@@ -35,7 +34,7 @@ public class ServiceController {
 
     @Operation(summary = "Update availability of a service")
     @PreAuthorize("hasRole('" + RoleConstants.SERVICE_PROVIDER_ROLE + "')")
-    @PatchMapping("/{serviceId}/availability")
+    @PatchMapping(ServiceControllerConstants.UPDATE_SERVICE_AVAILABILITY)
     public ResponseEntity<String> updateServiceAvailability(@PathVariable Long serviceId) {
         ServicesDTO updatedService =  servicesService.updateAvailability(serviceId);
         return ResponseEntity.ok("Service availability updated to " + updatedService.getAvailability());
@@ -44,7 +43,7 @@ public class ServiceController {
 
     @Operation(summary = "Get bookings for the current service provider")
     @PreAuthorize("hasRole('" + RoleConstants.SERVICE_PROVIDER_ROLE + "')")
-    @GetMapping("/bookings")
+    @GetMapping(ServiceControllerConstants.GET_SERVICE_BOOKINGS_URL)
     public ResponseEntity<Page<BookingDetailsResponse>> getBookings(Pageable pageable) {
         Page<BookingDetailsResponse> services = servicesService.getBookingsForServiceProvider(pageable);
         return ResponseEntity.ok(services);
@@ -52,13 +51,13 @@ public class ServiceController {
 
     @Operation(summary = "Accept or reject a booking request")
     @PreAuthorize("hasRole('" + RoleConstants.SERVICE_PROVIDER_ROLE + "')")
-    @PostMapping("/bookings/{bookingId}/status")
+    @PostMapping(ServiceControllerConstants.ACCEPT_OR_REJECT_SERVICE_BOOKING_URL)
     public Booking acceptOrRejectBooking(@PathVariable Long bookingId,@RequestParam BookingStatus status)   {
         return servicesService.respondToBookingRequests(bookingId, status);
     }
     @Operation(summary = "Cancel a booking as a service provider")
     @PreAuthorize("hasRole('" + RoleConstants.SERVICE_PROVIDER_ROLE + "')")
-    @PostMapping("/bookings/{bookingId}/cancel")
+    @PostMapping(ServiceControllerConstants.CANCEL_SERVICE_BOOKING_URL)
     public ResponseEntity<Void> cancelBooking(@PathVariable Long bookingId)  {
         servicesService.cancelBooking(bookingId);
         return ResponseEntity.ok().build();
