+++ src/main/java/com/example/cdr/eventsmanagementsystem/DTO/Booking/Request/BookingCancelRequest.java
@@ -3,6 +3,8 @@ package com.example.cdr.eventsmanagementsystem.DTO.Booking.Request;
 import lombok.Data;
 
 @Data
-public class CompletePaymentRequest {
-    private String paymentMethodId;
+public class BookingCancelRequest {
+    private Long bookingId;
+    
+    private String reason;  
 }
