+++ src/main/java/com/example/cdr/eventsmanagementsystem/DTO/Booking/Request/PaymentCompleteRequest.java
@@ -3,8 +3,6 @@ package com.example.cdr.eventsmanagementsystem.DTO.Booking.Request;
 import lombok.Data;
 
 @Data
-public class CancelBookingRequest {
-    private Long bookingId;
-    
-    private String reason;  
+public class PaymentCompleteRequest {
+    private String paymentMethodId;
 }
