+++ src/main/java/com/example/cdr/eventsmanagementsystem/Mapper/VenueMapper.java
@@ -7,7 +7,6 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.Named;
-
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
