+++ src/main/java/com/example/cdr/eventsmanagementsystem/Repository/VenueRepository.java
@@ -4,8 +4,6 @@ import com.example.cdr.eventsmanagementsystem.Model.Venue.Venue;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
 public interface VenueRepository extends JpaRepository<Venue,Long> {
 }
