+++ src/main/java/com/example/cdr/eventsmanagementsystem/Service/Auth/UserRoleHandler.java
@@ -2,6 +2,12 @@ package com.example.cdr.eventsmanagementsystem.Service.Auth;
 
 import com.example.cdr.eventsmanagementsystem.Model.User.BaseRoleEntity;
 
+/**
+ * Generic interface for handling operations on users with a specific role.
+ *
+ * @param <T> the type of user entity that extends BaseRoleEntity (e.g., Admin, Attendee, VenueProvider)
+ */
+
 public interface UserRoleHandler<T extends BaseRoleEntity> {
     boolean supports(String role);
     Class<T> getRoleClass(); // returns Admin.class, Attendee.class, etc.
