+++ src/main/java/com/example/cdr/eventsmanagementsystem/Service/Service/ServicesServiceInterface.java
@@ -1,13 +1,21 @@
 package com.example.cdr.eventsmanagementsystem.Service.Service;
 
+import com.example.cdr.eventsmanagementsystem.DTO.Booking.Response.BookingDetailsResponse;
 import com.example.cdr.eventsmanagementsystem.DTO.Service.ServicesDTO;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.Booking;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.BookingStatus;
-import java.nio.file.AccessDeniedException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+/**
+ * Service interface for managing services and their bookings.
+ */
+
 
 public interface ServicesServiceInterface {
     ServicesDTO addService(ServicesDTO dto);
-    ServicesDTO updateAvailability(Long serviceId) throws AccessDeniedException;
-    Booking respondToBookingRequests (Long bookingId, BookingStatus status) throws AccessDeniedException;
-    void cancelBooking(Long bookingId) throws AccessDeniedException;
+    ServicesDTO updateAvailability(Long serviceId);
+    Booking respondToBookingRequests (Long bookingId, BookingStatus status);
+    Page<BookingDetailsResponse> getBookingsForServiceProvider(Pageable pageable);
+    void cancelBooking(Long bookingId);
 }
